requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
flake8==4.0.1
idna==3.3
isort==5.10.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
requests==2.28.0
tabulate==0.8.9
urllib3==1.26.9
